backend/requirements.txt
pyserial==3.5
websockets==12.0
brainflow>=5.10.0
numpy>=1.24.0